requirements.txt
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smart-open==3.0.0
smmap==4.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==3.0.5
spacy-legacy==3.0.2
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
srsly==2.4.1
statsmodels==0.11.0
stop-words==2018.7.23
stopwords==1.0.0
streamlit==0.82.0
sympy==1.5.1
tables==3.6.1
tablib==2.0.0
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
thinc==8.0.3
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
utils==1.0.1
validators==0.18.2
virtualenv==20.2.1
wasabi==0.8.2
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xgboost==1.2.1
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==1.0.0
zipp==2.2.0
streamlit==0.82.0
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
apipkg==1.5
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.0
appscript==1.0.1
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.0
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
black==21.5b1
bleach==3.1.0
blinker==1.4
blis==0.7.4
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.2
Brotli==1.0.9
cachetools==4.2.2
catalogue==2.0.3
category-encoders==2.2.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
conda==4.9.1
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
cytoolz==0.10.1
dash==1.17.0
dash-core-components==1.13.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
dask==2.11.0
de-core-news-md==3.0.0
de-core-news-sm==3.0.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
distributed==2.11.0
docopt==0.6.2
docutils==0.16
en-core-web-sm==3.0.0
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.7.1
fastcache==1.1.0
filelock==3.0.12
filterpy==1.4.5
flake8==3.7.9
Flask==1.1.1
Flask-Compress==1.8.0
fsspec==0.6.2
future==0.18.2
gensim==4.0.1
gevent==1.4.0
gitdb==4.0.7
GitPython==3.1.17
glob2==0.7
gmpy2==2.0.8
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-speech==2.3.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphviz==0.14.2
greenlet==0.4.15
grpcio==1.37.1
h5py==2.10.0
HeapDict==1.0.1
helper-functions==2.0.11
html5lib==1.0.1
httplib2==0.19.1
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imagesize==1.2.0
imbalanced-learn==0.7.0
importlib-metadata==1.5.0
imutils==0.5.3
iniconfig==1.0.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
libcst==0.3.18
lief==0.9.0
lightgbm==3.1.1
linearmodels==4.24
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlens==0.2.3
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.5
mypy-extensions==0.4.3
mysqlclient==2.0.3
natsort==7.0.1
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauth2client==4.1.3
olefile==0.46
opencv-python==4.4.0.46
openpyxl==2.4.11
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathspec==0.8.1
pathtools==0.1.2
pathy==0.4.0
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plotly==4.13.0
pluggy==0.13.1
ply==3.11
preshed==3.0.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
property-cached==1.6.4
propscore==0.0.5
proto-plus==1.18.1
protobuf==3.15.8
psutil==5.6.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.7.3
pydeck==0.6.2
pydocstyle==4.0.1
pydotplus==2.0.2
pyflakes==2.1.1
Pygments==2.5.2
pyhdfe==0.1.0
pylint==2.4.4
PyMySQL==0.10.1
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadr==0.4.0
pyrsistent==0.15.7
PySocks==1.7.1
pytest==6.0.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-forked==1.3.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pytest-xdist==2.0.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyyaml==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
records==0.5.3
regex==2021.4.4
requests==2.22.0
retrying==1.3.3
rope==0.16.0
rpy2==3.4.3
rsa==4.7.2
Rtree==0.9.3
ruamel-yaml==0.15.87
schedule==0.6.0
scikit-image==0.16.2
